cv_checker/requirements.txt
pdfplumber 
python-docx 
spacy 
streamlit
